refactor: move the Add To Space modal into a separate spaces component
